requirements.txt
discord.py[voice]>=2.7.1
PyNaCl
httpx
openai
pyyaml
apscheduler>=3.10.0
ollama
python-dotenv

# Web portal dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
typing_extensions
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
rich
requests
yfinance
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
azure-cognitiveservices-speech>=1.40.0